requirements.txt
alembic==1.8.1
anyio==3.6.1
apturl==0.5.2
attrs==22.1.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
cleo==0.8.1
Click==7.0
clikit==0.6.2
colorama==0.4.3
command-not-found==0.3
crashtest==0.3.1
cryptography==37.0.4
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
duplicity==0.8.12.0
ecdsa==0.18.0
email-validator==1.2.1
entrypoints==0.3
fastapi==0.79.0
fasteners==0.14.1
filelock==3.8.0
future==0.18.2
greenlet==1.1.2
h11==0.9.0
html5lib==1.1
httplib2==0.14.0
httptools==0.1.2
idna==2.8
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
jeepney==0.8.0
keyring==23.8.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
msgpack==1.0.4
mysql-connector-python==8.0.30
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.6.0
passlib==1.7.4
pastel==0.2.1
pexpect==4.8.0
Pillow==7.0.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.1.14
poetry-core==1.0.8
protobuf==3.20.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.9.1
PyGObject==3.36.0
PyJWT==1.7.1
pylev==1.4.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
pytest==7.1.2
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
systemd-python==234
tomli==2.0.1
tomlkit==0.11.4
typing-extensions==4.3.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.11.3
uvloop==0.16.0
virtualenv==20.16.3
wadllib==1.3.3
webencodings==0.5.1
websockets==8.1
wsproto==0.15.0
xkit==0.0.0
zipp==3.8.1
